moved sync into its own service folder
