+++ plistsync/services/sync/__init__.py
@@ -1,6 +1,6 @@
 """Synchronisation primitives for cross-service playlist management."""
 
-from plistsync.core.sync.playlist import SyncedPlaylist, SyncedPlaylistID
+from plistsync.services.sync.playlist import SyncedPlaylist, SyncedPlaylistID
 
 __all__ = [
     "SyncedPlaylist",
